+++ spec/lib/aixm/refinements_spec.rb
@@ -34,42 +34,6 @@ describe AIXM::Refinements do
         _([nil].to_digest).must_equal "cf83e135"
       end
     end
-
-    describe :find do
-      subject do
-        [AIXM::Factory.dme, AIXM::Factory.ndb, AIXM::Factory.vor]
-      end
-
-      it "finds elements by exact class" do
-        _(subject.find(is_a: AIXM::Feature::NavigationalAid::DME).count).must_equal 1
-      end
-
-      it "finds elements by fuzzy class" do
-        _(subject.find(kind_of: AIXM::Feature).count).must_equal 3
-      end
-
-      it "finds elements by attribute" do
-        _(subject.find(id: 'MMM').count).must_equal 1
-      end
-
-      it "finds elements by class and attributes" do
-        _(subject.find(is_a: AIXM::Feature::NavigationalAid::DME, id: 'MMM', name: 'DME NAVAID').count).must_equal 1
-      end
-
-      it "returns empty array if nothing is found" do
-        _(subject.find(is_a: Array).count).must_equal 0
-      end
-    end
-
-    describe :duplicates do
-      it "returns an array containing the duplicates" do
-        _(%w(a b c b d e d f b).duplicates).must_equal %w(b d)
-      end
-
-      it "returns an empty array if no duplicates are present" do
-        _(%w(a b c d e f g h i).duplicates).must_equal []
-      end
-    end
   end
 
   context Float do